requirements.txt
# ─── Django Core ──────────────────────────────────────────────────────────────
Django>=4.2,<5.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# ─── REST API ─────────────────────────────────────────────────────────────────
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# ─── Static Files ─────────────────────────────────────────────────────────────
whitenoise>=6.6.0

# ─── Image Processing ─────────────────────────────────────────────────────────
Pillow>=10.0.0

# ─── Payment ──────────────────────────────────────────────────────────────────
razorpay>=1.4.1

# ─── HTTP Requests ────────────────────────────────────────────────────────────
requests>=2.31.0

# ─── Utils ────────────────────────────────────────────────────────────────────
python-decouple>=3.8
qrcode[pil]>=7.4.2

# ─── AWS ──────────────────────────────────────────────────────────────────────
boto3>=1.34.0
django-ses>=3.5.0
dj-database-url>=2.1.0

# ─── RAG Chatbot ──────────────────────────────────────────────────────────────
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1